User/Admin ProductStackView를 개별 타입화, View 구성 메소드 이전
